extract intervals.ts as a separate module
